+++ example/tests/runner.php
@@ -6,6 +6,7 @@
  */
 
 load_config('config/local');
+run_suite('click');
 run_suite('attributes');
 run_suite('suite');
 run_suite('suite_dev_only');